moved game Ninja to the games package
